feat: add 'purchasing', 'winning' domains and reorganize classes
