+++ src/main/java/lotto/winning/WinningChecker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.winning;
 
 public class WinningChecker {
 }
